# Improve performance with `recursive_copy`