@ZacLN 2020-05-30
Thank this looks fantastic!

With regards to the timings and how this translates to real world responsiveness - the copying of the `SymbolServer.stdlibs` store is a once per session operation so this change will benefit the startup time for the language server. The intention is to fiddle with the SymbolServer storage objects' structure as little as possible so maintaining the `recursive_copy` functions isn't an issue.

The following snippet would time the indexing of a package:
```julia
using SymbolServer
using Packages,You,Want,To,Index
env = SymbolServer.getenvtree() #This creates a tree of all modules within the current session, including submodules
@time SymbolServer.symbols(env) # index everything
@time SymbolServer.symbols(env, somemodule) # index a single module
```
For realistic timings I think you  need to restart the Julia session between runs, though for checking the impact of changes it's probably fine to not. If I remember correctly a significant share of the total time is spent introspecting methods