@timholy 2020-05-30
Thanks for the great tips, @ZacLN! The next set of changes I made were orthogonal to these, so I put them in a separate PR (#160). As far as I am concerned this PR is all set, if you want it. (It probably doesn't affect performance in real-world usage all that significantly, esp. compared to #160, so whatever you choose will be fine with me.)