@ZacLN 2020-05-30
Definitely one to merge! Reductions in startup time are always worth double, right?